+++ CHANGES.md
@@ -1,5 +1,10 @@
 # CHANGES
 
+## 4.2.0
+
+* Add threadpool to `bam_stats` and `diff_bams`. Fixes #18, Fixes #19
+* bwa_mem.pl can use threads for `bam_stats`
+
 ## 4.1.4
 
 * Ensure `sentinel` calls exit with non-zero exit code. Fixes #16